requirements.txt
asgiref==3.4.1
attrs==21.4.0
black==21.9b0
CacheControl==0.12.11
cachetools==5.0.0
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.6
click==8.0.1
cloudinary==1.26.0
crypto==1.4.1
cryptography==37.0.2
Deprecated==1.2.13
dj-database-url==0.5.0
Django==4.0.5
django-cors-headers==3.8.0
django-heroku==0.0.0
djangorestframework==3.12.4
dnspython==1.16.0
firebase==3.0.1
firebase-admin==5.2.0
gcloud==0.18.3
google-api-core==2.7.3
google-api-python-client==2.47.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-cloud-core==2.3.0
google-cloud-firestore==2.4.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpcio==1.46.0
grpcio-status==1.46.0
gunicorn==20.1.0
httplib2==0.20.4
idna==3.2
iniconfig==1.1.1
install==1.3.5
jwcrypto==1.2
msgpack==1.0.3
mypy-extensions==0.4.3
Naked==0.1.31
numpy==1.22.4
oauth2client==4.1.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.0
pluggy==1.0.0
proto-plus==1.20.3
protobuf==3.20.1
psycopg2-binary==2.9.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pymongo==3.12.0
pyparsing==3.0.8
pytest==7.1.2
python-dotenv==0.19.0
python-jwt==3.3.2
pytz==2021.3
PyYAML==6.0
regex==2021.9.30
requests==2.26.0
rsa==4.8
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.2
sseclient==0.0.27
tomli==1.2.1
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
wrapt==1.14.1
